[Admin] Extract an admin-only ProductType
