--- src/Sylius/Bundle/AdminBundle/Form/Extension/ProductTypeExtension.php
@@ -1,53 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Sylius Sp. z o.o.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Sylius\Bundle\AdminBundle\Form\Extension;
-
-use Sylius\Bundle\AdminBundle\Form\Type\ProductAssociationsType;
-use Sylius\Bundle\CoreBundle\Form\Type\Product\ProductImageType;
-use Sylius\Bundle\ProductBundle\Form\Type\ProductType;
-use Symfony\Component\Form\AbstractTypeExtension;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\UX\LiveComponent\Form\Type\LiveCollectionType;
-
-final class ProductTypeExtension extends AbstractTypeExtension
-{
-    /**
-     * @param array<string, mixed> $options
-     */
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder
-            ->add('images', LiveCollectionType::class, [
-                'entry_type' => ProductImageType::class,
-                'entry_options' => ['product' => $options['data']],
-                'allow_add' => true,
-                'allow_delete' => true,
-                'by_reference' => false,
-                'label' => 'sylius.form.product.images',
-                'block_name' => 'entry',
-            ])
-            ->add('associations', ProductAssociationsType::class, [
-                'label' => false,
-            ])
-        ;
-    }
-
-    /**
-     * @return iterable<class-string>
-     */
-    public static function getExtendedTypes(): iterable
-    {
-        return [ProductType::class];
-    }
-}
